pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pcm"
version = "0.1.0"
description = "Color management course materials"
requires-python = ">=3.12"
dependencies = [
  "ipykernel",
  "jupyter",
  "numpy",
  "plotly",
  "dash",
  "kaleido",
  "matplotlib",
  "scipy",
  "colour-science[optional]",
  "types-opencolorio"
]

[project.optional-dependencies]
dev = [
  "nbstripout",
  "pre-commit",
]

[tool.setuptools.packages.find]
where = ["src"]
